Iterate through Span instead of List. It will def help with big Pipes

### Before
|             Method |       Mean |     Error |    StdDev |   Gen0 |   Gen1 | Allocated |
|------------------- |-----------:|----------:|----------:|-------:|-------:|----------:|
|          EmptyPipe |   9.112 ns | 0.0063 ns | 0.0049 ns |      - |      - |         - |
|          RetryPipe |  71.486 ns | 0.3123 ns | 0.2768 ns | 0.0305 |      - |      64 B |
|    ConcurrencyPipe |  85.402 ns | 0.7407 ns | 0.6185 ns |      - |      - |         - |
|         DoublePipe | 139.382 ns | 1.7267 ns | 1.5307 ns | 0.0305 | 0.0002 |      64 B |
|       DispatchPipe | 155.674 ns | 3.1004 ns | 3.3174 ns | 0.0956 |      - |     200 B |
| DoubleDispatchPipe | 244.221 ns | 1.1462 ns | 1.0722 ns | 0.1602 |      - |     336 B |
| TripleDispatchPipe | 260.601 ns | 2.8179 ns | 2.6359 ns | 0.1645 |      - |     344 B |

```
// * Hints *
Outliers
  SendBenchmark.EmptyPipe: .NET 6.0       -> 3 outliers were removed (10.32 ns..10.48 ns)
  SendBenchmark.RetryPipe: .NET 6.0       -> 1 outlier  was  removed (74.45 ns)
  SendBenchmark.ConcurrencyPipe: .NET 6.0 -> 2 outliers were removed (89.51 ns, 94.70 ns)
  SendBenchmark.DoublePipe: .NET 6.0      -> 1 outlier  was  removed (147.39 ns)

// * Legends *
  Mean      : Arithmetic mean of all measurements
  Error     : Half of 99.9% confidence interval
  StdDev    : Standard deviation of all measurements
  Gen0      : GC Generation 0 collects per 1000 operations
  Gen1      : GC Generation 1 collects per 1000 operations
  Allocated : Allocated memory per single operation (managed only, inclusive, 1KB = 1024B)
  1 ns      : 1 Nanosecond (0.000000001 sec)
```

### After
|             Method |       Mean |     Error |    StdDev |   Gen0 | Allocated |
|------------------- |-----------:|----------:|----------:|-------:|----------:|
|          EmptyPipe |   8.834 ns | 0.0466 ns | 0.0413 ns |      - |         - |
|          RetryPipe |  70.291 ns | 0.4289 ns | 0.3802 ns | 0.0305 |      64 B |
|    ConcurrencyPipe |  83.644 ns | 0.3903 ns | 0.3460 ns |      - |         - |
|         DoublePipe | 137.754 ns | 0.2401 ns | 0.2128 ns | 0.0305 |      64 B |
|       DispatchPipe | 148.460 ns | 0.1048 ns | 0.0929 ns | 0.0956 |     200 B |
| DoubleDispatchPipe | 239.103 ns | 0.2064 ns | 0.1930 ns | 0.1602 |     336 B |
| TripleDispatchPipe | 254.960 ns | 0.2282 ns | 0.2023 ns | 0.1645 |     344 B |

```
// * Hints *
Outliers
  SendBenchmark.EmptyPipe: .NET 6.0          -> 1 outlier  was  removed (10.16 ns)
  SendBenchmark.RetryPipe: .NET 6.0          -> 1 outlier  was  removed (75.54 ns)
  SendBenchmark.ConcurrencyPipe: .NET 6.0    -> 1 outlier  was  removed (87.73 ns)
  SendBenchmark.DoublePipe: .NET 6.0         -> 1 outlier  was  removed (140.51 ns)
  SendBenchmark.DispatchPipe: .NET 6.0       -> 1 outlier  was  removed (151.16 ns)
  SendBenchmark.TripleDispatchPipe: .NET 6.0 -> 1 outlier  was  removed (258.42 ns)

// * Legends *
  Mean      : Arithmetic mean of all measurements
  Error     : Half of 99.9% confidence interval
  StdDev    : Standard deviation of all measurements
  Gen0      : GC Generation 0 collects per 1000 operations
  Allocated : Allocated memory per single operation (managed only, inclusive, 1KB = 1024B)
  1 ns      : 1 Nanosecond (0.000000001 sec)
```